# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: "Spotplay API" 
  version: 1.0.0
  description: "Una plataforma de música"
  contact:
    email: jealtzo26@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:4000/api/v1
    description: "Localhost"
...
components:
  schemas: 
...
    Playlist:
      type: object
      properties:
        name:
          type: string
        idOwner: 
          type: integer
        description:
          type: string
    PlaylistResponse:
      type: object
      properties:
        _id:
          type: integer
        _name:
          type: string
        _idOwner: 
          type: integer
        _description:
          type: string
...
paths:
...
  /playlist:
    post:
      tags:
      - Playlist
      summary: Crear una playlist
      description: Se creará una playlist para utilizarla en la aplicación      
      requestBody: 
        description: Propiedades para crear una playlist 
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Playlist'
      responses:
        201:
          description: "Registro exitoso"
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistResponse'
        400:
          description: Bad request
          content: {}
          
    get:
      tags:
      - Playlist
      summary: Todos los playlists
      description: Retorna todos los playlists de la plataforma      
      parameters:
        - in: query
          name: idOwner          
          schema:
            type: string
      responses:
        200:
          description: Array de todos las playlists
          content:
            application/json:
              schema:
                type: array
                items:
                    $ref: '#/components/schemas/PlaylistResponse'
        400:
          description: Bad request
          content: {}
  /playlist/{id}:
    get:
      tags:
      - Playlist
      summary: Playlist por id
      description: Retorna una playlist con un id específico
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            
          description: ID del playlist
      responses:
        200:
          description: Playlist solicitado
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/PlaylistResponse'
        400:
          description: Bad request
          content: {}
    put:
      tags:
      - Playlist
      summary: Actualizar playlist
      description: Actualiza una playlist en específico
      requestBody: 
        description: Al menos uno de todos los atributos
        content: 
          application/json:
            schema:
                type: object
                properties:
                  _name:
                    type: string
                  _description:
                    type: string                  
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            
          description: ID del playlist
      responses:
        200:
          description: Playlist solicitado
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/PlaylistResponse'
        400:
          description: Bad request
          content: {}
          
    delete:
      tags:
      - Playlist
      summary: Eliminar playlist
      description: Elimina una playlist en específico
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            
          description: ID de la playlist
      responses:
        200:
          description: Eliminado con éxito
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/PlaylistResponse'
        400:
          description: Bad request
          content: {}
  /playlist/{id}/{idSong}:
    post:
      tags:
      - Playlist
      summary: Agregar una canción
      description: Agregar una canción (idSong) a una playlist (id)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            
          description: ID de la playlist
        - in: path
          name: idSong
          required: true
          schema:
            type: string            
          description: ID de la canción
      responses:
        201:
          description: Canción agregada
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/PlaylistResponse'
        400:
          description: Bad request
          content: {}
        

# routes for role
...